lexer: Fix identifier handling
For some reason (I'm too groggy to sort that one out) the way I
constructed string broke everything. Fix that. It wasn't needed anyway
(since it is already appended in the while loop).

P.S. I think, I figured it out. probably it was uninitialized or 0,
which is indeed made all comparisons false.