AST: Add TypeDeclaration
Solution to the problem: InstanceDeclaration derives just from
Declaration, while Prototype and Class derive from TypeDeclaration.
This is not entirely correct semantically (prototype is not a new type,
but rather just a template for an instance, or at least it seemes to be
so), but allows for instances to share properties without duplicating
code.