lexer: Add list of token macro-definitions
I am not sure what to call it, but I figured, it would be simpler to use
macro, then assemble that by hand or use external tools.
It also makes it easier to add new TokenKinds, since they are now in one
place and I won't have to change them everywhere.