lexer: Store characters composing token in token
I didn't store characters for operators, since these weren't needed to
parse them (unlike strings, numbers and identifiers), but these are good
to have for error messages. This is a temporary hack, which won't be
merged into master, so I dunno why I am writing this long explanation.
The point of this hack is, that I don't want to build up the string,
so instead I am using the 'spellToken()' thing.
This will get fixed when I implement SourceBuffer.