parser: implement parse_unary_expression
I want to get into the code generation ASAP so that I can I have a base
all the way through. I will implement simple expressions and skip stuff
like classes, variables and loops for now.

I decided to redesign the AST and use std::variant instead of
inheritance. I just find it easier to work with variants that with
classic OO code. And this will be easier to translate into awscript
later, because it will look like

    define expression = unary_expression | binary_expression | value;

in awscript