ast: rework statements
I replaced them with a variant, by analogy with the expression. Now
there's an extra allocation, but I don't know how to deal with it yet.
One possibility is to make all expression parsers return stack-allocated
values.
In awscript this would be a lot simpler, because ( stmt | expr )* and (
stmt* | expr* ) are the same .