Add empty_statement
I need something that is not nullptr for constructs like 'if x > 0;'.
nullptr indicates an error, which is not what is expected in this case.
Also I might want to put attributes on empty statements in the future.
I could use an empty statement_block, but it feels less natural.