Implement if_expression
I am not yet sure if I want to go rust way and turn a statement block
into an expression. So I'm doing this in two steps: first, make the
if_expression; second, do variations on that (like turning the statement
block into an expression or extending the if_expression itself).