Implement rudimentary type system
In its current form it is so rudimentary that I should even call it a
type system.

So this is just a standin for a real type system that allows me to have
primitive types in my code until I implement the real type system.

At the very least I need aliases to avoid hardcoding a bunch of ints.
Another thing that requires immediate attention is a basic type checker
with support for type conversions (to support integer_literal and
others)
And of course the mapping of type* to llvm::Type* instead of looking at
strings