Allow omitting braces for single-line functions
Now functions that are made out of a single return statement don't need
braces:
    function f() : int
        return 1;

There's also disalbed code for implicit return:
    function f() : int
        1;

I find this syntax not quite clear, and there's a bit of inconsistency
with "lambda" functions (which aren't implemented yet, but I ran it in
my head).

I may revisit that later, and enable that at least for `void` functions.