ast: convert declaration to a variant
It kinda stood out since I converted expressions and statements into
variants. Overall, it allows to reduce the amount of code across the board.
There are a few problems with it, like a need to make error_t more
specific, and a need to move some things into a .cpp file.
However, overall, I believe it allows to simplify the code.

It also reveals deficiencies the current design: symtab and scope are
redundant. Scope can be reworked and used in the future to optimize
lookups, but symtab is going to be replaced by a module.