Add crude type inference
It is crude and inefficient, but appears to work (I'm not quite sure, as
I didn't write tests and only tested it manually).

The algorithm implemented here isn't quite what it is supposed to be.
Roughly, it is supposed to work like this
1. Given a constraint T', propagate it down the tree
2. Walk up the tree, inferring the type for each node

Thus far, there is no notion of constraints in the code, and it doesn't
assign type to each expression. But it does what it is written for - it
gives each numeric_literal a type, so that I can use types other than int.