backend_llvm: change strign literal linkage
Changed the linkage to private (not sure yet if this will work as I
indend it to work, i.e. if literals in different modules will be
collapsed)

Added the unnamed_addr mark. Doesn't appear to affect anything, but it
signifies that multiple of the same string can be collapsed together.

Also enabled PIC, because ld complains about constant strings in .rodata