parser: implement inline modules
Inline modules are modules that are nested within their parent module.
For now it just parses the body and returns an not-implemented-yet
error. I adapted the code from the WIP branch but I changed it so it no
longer adds extra state to the parser (it is stored in the call stack
instead).
Since writing `module x { module x; }` doesn't make much sense, I also
introduce the concept of a `decl_context` that allows to forbid certain
types of declarations in certain context. I also applied it to the
foreign blocks.