semantic: add resolve_dependencies()
Goes over AST and lists all dependencies. For now I do it in a form
of std::map but if it proves to be a bottleneck, I'll revisit it later.

The idea is as follows:
- process the AST and extract all dependencies (this step)
- locate the dependencies and parse them to get the AST
- repeat the process for all the dependencies
- keep doing this until all dependencies have been resolved
- dependencies that were already parsed aren't re-parsed again

As a first step, all of this will be done in a single process, single
thread. Later I'll parallelize the algorithm and much later I'll split
it into several processes but that requires a stable file format for
modules as a prerequisite.

TODO: Perhaps "resolve_dependencies" isn't quite accurate name