semantic: refactor dependency_resolver
First, I moved dependency_scanner into another header because there are
going to be more classes.
Second, I decided to ditch the map after all and replace it with just a
vector. I realized that I don't actually need a map at this point and a
linear vector is easier to work with than a tree.